# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Fully control your smart home's Lamps using your mobile app"
  version: "1.0.0"
  title: "Smart bulb API"
host: "ibtikar-gl-bulb.appspot.com"
basePath: "/api"
schemes:
- "https"
- "http"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /users:
    post:
      tags:
      - "USERS"
      summary: "add new user / create account"
      description: "This function allows user to create new account, accounts are\
        \ unique\nby email address, so one user can't create more than one account\
        \ using \nthe same email address.\n"
      operationId: "usersPOST"
      parameters:
      - name: "x-bulb-api-key"
        in: "header"
        description: "secret hash SHA1 of 'ibtikar+glinty+smart+bulb' is '5932e7543873bbfa2d19c32649092c73d0bdcb1f'"
        required: true
        type: "string"
      - in: "body"
        name: "user object"
        description: "user object"
        required: true
        schema:
          $ref: "#/definitions/userObject"
      responses:
        200:
          description: "user object."
          schema:
            $ref: "#/definitions/User"
        401:
          description: "Data Descriped Error."
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Authorized Request / Not Found."
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "General Error Handler."
          schema:
            $ref: "#/definitions/Error"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "USERS"
  /users/{USER_ID}:
    get:
      tags:
      - "USERS"
      summary: "Get user info"
      description: "This returns you all info related to a user by passing userID\
        \ to the\nAPI.\n"
      operationId: "usersUSER_IDGET"
      parameters:
      - name: "USER_ID"
        in: "path"
        description: "user id to retrieve info about."
        required: true
        type: "number"
        format: "int32"
      - name: "x-bulb-api-key"
        in: "header"
        description: "secret hash SHA1 of 'ibtikar+glinty+smart+bulb' is '5932e7543873bbfa2d19c32649092c73d0bdcb1f'"
        required: true
        type: "string"
      responses:
        200:
          description: "User object."
          schema:
            $ref: "#/definitions/User"
        401:
          description: "Data Descriped Error."
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Authorized Request / Not Found."
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "General Error Handler."
          schema:
            $ref: "#/definitions/Error"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "USERS"
    put:
      tags:
      - "USERS"
      summary: "Update user account."
      description: "This function allows user to update his/her account, accounts\
        \ are unique by email address.\n"
      operationId: "usersUSER_IDPUT"
      parameters:
      - name: "USER_ID"
        in: "path"
        description: "userid to delete"
        required: true
        type: "number"
        format: "int32"
      - name: "x-bulb-api-key"
        in: "header"
        description: "secret hash SHA1 of 'ibtikar+glinty+smart+bulb' is '5932e7543873bbfa2d19c32649092c73d0bdcb1f'"
        required: true
        type: "string"
      - in: "body"
        name: "user object"
        description: "user object"
        required: true
        schema:
          $ref: "#/definitions/userObject"
      responses:
        200:
          description: "user object."
          schema:
            $ref: "#/definitions/User"
        401:
          description: "Data Descriped Error."
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Authorized Request / Not Found."
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "General Error Handler."
          schema:
            $ref: "#/definitions/Error"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "USERS"
    delete:
      tags:
      - "USERS"
      summary: "Delete user account."
      description: "This function allows user to delete his/her account, accounts\
        \ are unique by email address.\n"
      operationId: "usersUSER_IDDELETE"
      parameters:
      - name: "USER_ID"
        in: "path"
        description: "userid to delete"
        required: true
        type: "number"
        format: "int32"
      - name: "x-bulb-api-key"
        in: "header"
        description: "secret hash SHA1 of 'ibtikar+glinty+smart+bulb' is '5932e7543873bbfa2d19c32649092c73d0bdcb1f'"
        required: true
        type: "string"
      responses:
        200:
          description: "done"
        401:
          description: "Data Descriped Error."
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Authorized Request / Not Found."
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "General Error Handler."
          schema:
            $ref: "#/definitions/Error"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "USERS"
definitions:
  User:
    type: "object"
    properties:
      id:
        type: "number"
        description: "userid is unique. Used for all user requests."
      username:
        type: "string"
        description: "username used for registeration."
      email:
        type: "string"
        description: "user email used for registeration / Unique."
      password:
        type: "string"
        description: "password for user."
      birth_date:
        type: "string"
        description: "user birthdate in date formate"
      latitude:
        type: "number"
        description: "user latitude for mapping his/her place"
      longitude:
        type: "number"
        description: "user longitude for mapping his/her place"
      created_at:
        type: "string"
        description: "DateString that indicates when this profile created"
      modified_at:
        type: "string"
        description: "DateString that indicates when this profile modified"
  Error:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      message:
        type: "string"
  userObject:
    type: "object"
    required:
    - "pass"
    - "uem"
    - "un"
    properties:
      un:
        type: "string"
        description: "username"
      uem:
        type: "string"
        description: "user email address."
      pass:
        type: "string"
        format: "password"
        description: "user account password MD5ed"
      bd:
        type: "string"
        format: "date-time"
        description: "user birth date in string formate"
      lon:
        type: "number"
        format: "double"
        description: "user location longitude"
      lat:
        type: "number"
        format: "double"
        description: "user location latitude"
    description: "user object"
...